chore: move useEvmosBalance to evmos wallet
