+++ apps/governance/package.json
@@ -1,6 +1,6 @@
 {
   "name": "governance-page",
-  "version": "1.0.15",
+  "version": "1.0.16",
   "private": true,
   "scripts": {
     "dev": "next dev --port 3001",
