+++ apps/mission/src/internal/types.ts
@@ -144,13 +144,6 @@ export type ERC20BalanceResponse = {
   balance: ERC20Element[];
 };
 
-export type BalanceResponse = {
-  balance: {
-    amount: string;
-    denom: string;
-  };
-};
-
 type AmountProposal = {
   denom: string;
   amount: string;
