+++ packages/evmos-wallet/package.json
@@ -1,6 +1,6 @@
 {
   "name": "evmos-wallet",
-  "version": "1.0.10",
+  "version": "1.0.11",
   "sideEffects": [
     "**/*.css"
   ],
