+++ packages/evmos-wallet/src/internal/wallet/functionality/fetch.ts
@@ -1,6 +1,14 @@
 // Copyright Tharsis Labs Ltd.(Evmos)
 // SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/apps/blob/main/LICENSE)
 
+import {
+  EVMOS_BACKEND,
+  EVMOS_MINIMAL_COIN_DENOM,
+  EVMOS_SYMBOL,
+} from "./networkConfig";
+
+import { BalanceResponse } from "./types";
+
 export async function fetchWithTimeout(
   resource: string,
   options: RequestInit & { timeout?: number } = {}
@@ -15,3 +23,13 @@ export async function fetchWithTimeout(
   clearTimeout(id);
   return response;
 }
+
+export const getEvmosBalance = async (address: string) => {
+  if (address === "" || address == undefined || address == null) {
+    return { balance: { denom: "", amount: "" } };
+  }
+  const res = await fetch(
+    `${EVMOS_BACKEND}/BalanceByDenom/${EVMOS_SYMBOL}/${address}/${EVMOS_MINIMAL_COIN_DENOM}`
+  );
+  return res.json() as Promise<BalanceResponse>;
+};
